+++ modules/colormatch.py
@@ -22,6 +22,22 @@ from torchvision.transforms.v2 import ToTensor, ToPILImage
 
 
 
+# ========= pil2tensor ========= #
+def pil2tensor(image):
+    print("[DICKSON-NODES] pil2tensor")
+    return torch.from_numpy(np.array(image).astype(np.float32) / 255.0).unsqueeze(0)
+
+
+
+# ========= tensor2pil ========= #
+def tensor2pil(image):
+    print("[DICKSON-NODES] tensor2pil")
+    return Image.fromarray(
+        np.clip(255.0 * image.cpu().numpy().squeeze(), 0, 255).astype(np.uint8)
+    )
+
+
+
 # Main function for AdaIN color matching
 def adain_color_match(target: Image, source: Image):
     print("[DICKSON-NODES] adain_color_match")
